Consistency
Move the Theme enum to the root for 'easier access'

Change the AbstractTheme into a normal class - we don't need it to be abstract
